TerrainType: replace ID with TerrainID where equivalent
First there was only TerrainID, the primary key. Then due to the non-"ID" name a new column was auto-generated. At some point someone noticed that and the IDs got made equivalent to TerrainID. It is unknown whether the column is now called TerrainID or ID, but TerrainID values are consistent since Vanilla with *this* naming scheme.